Move utility classes to utilities subfolder - fix test classes
